--- backend/src/Users/index.ts
@@ -1,3 +1,0 @@
-import { UsersModule } from './users.module';
-
-export { UsersModule };
