+++ backend/src/chatRoomMembers/index.ts
@@ -1,0 +1,3 @@
+import { ChatRoomMembersModule } from './chatRoomMembers.module';
+
+export { ChatRoomMembersModule };
