+++ backend/src/common/dto/chatRoom.dto.ts
@@ -2,5 +2,5 @@ export interface ChatRoomDto {
   id?: number;
   name: string;
   created_by: number;
-  isPublic: boolean;
+  is_public: boolean;
 }
