+++ frontend/src/Chat/components/ChatList/AlertElement.tsx
@@ -1,0 +1,20 @@
+import { useState } from 'react'
+import { Alert } from 'react-bootstrap'
+
+export const AlertElement = (props: { kicked: boolean }): JSX.Element => {
+  const [show, setShow] = useState<boolean>(props.kicked)
+
+  return show ? (
+    <Alert
+      variant="danger"
+      onClose={() => {
+        setShow(false)
+      }}
+      dismissible
+    >
+      <Alert.Heading>You are kicked from the Chat room! </Alert.Heading>
+    </Alert>
+  ) : (
+    <></>
+  )
+}
