+++ frontend/src/Chat/index.tsx
@@ -1,4 +1,4 @@
-import { Chat } from './components/Chat'
-import { ChatList } from './components/ChatList'
+import { Chat } from './components/Chat/Chat'
+import { ChatList } from './components/ChatList/ChatList'
 
 export { Chat, ChatList }
