+++ swagger/api.yml
@@ -15,6 +15,8 @@ tags:
     description: チャットルーム
   - name:  chat-mute-user
     description: チャット機能でのミュートユーザ情報
+  - name: chatRoomMembers
+    description: チャットルームの参加者情報
 
 paths:
   /health:
@@ -250,7 +252,51 @@ paths:
                 type: array
                 items:
                   $ref: '#/components/schemas/chat-mute-user'
-
+  /chatRoomMembers:
+    get:
+      tags:
+        - chatRoomMembers
+      summary: get chatRoomMembers
+      description: チャット機能でのルームの参加者情報を取得
+      responses:
+        200:
+          description: 成功
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/chatRoomMember'
+    post:
+      summary: update chatRoomMembers
+      tags:
+        - chatRoomMembers
+      description: チャット機能でのルームの参加者情報を更新
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/chatRoomMember'
+      responses:
+        201:
+          description: 成功
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/chatRoomMember'
+    delete:
+      summary: delete chatRoomMembers
+      tags:
+        - chatRoomMembers
+      description: チャット機能でのルームの参加者情報を削除
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/chatRoomMemberPK'
+      responses:
+        200:
+          description: 成功
 components:
   schemas:
     health:
@@ -346,7 +392,7 @@ components:
           type: number
           example: 2
         mute_until:
-          type: timestamp
+          type: string
           example: "2024-01-31T14:59:00.000Z"
     chat-mute-user-pk:
       type: object
@@ -357,3 +403,24 @@ components:
         mutedUserId:
           type: number
           example: 2
+    chatRoomMember:
+      type: object
+      properties:
+        chatRoomId:
+          type: number
+          example: 1
+        userId:
+          type: number
+          example: 1
+        isBanned:
+          type: boolean
+          example: false
+    chatRoomMemberPK:
+      type: object
+      properties:
+        chatRoomId:
+          type: number
+          example: 1
+        userId:
+          type: number
+          example: 1
